update to glumpy master head

+++ examples/interpolations.py
@@ -1,7 +1,5 @@
-#! /usr/bin/env python
-# -*- coding: utf-8 -*-
 # -----------------------------------------------------------------------------
-# Copyright (c) 2014, Nicolas P. Rougier. All Rights Reserved.
+# Copyright (c) 2009-2016 Nicolas P. Rougier. All rights reserved.
 # Distributed under the (new) BSD License.
 # -----------------------------------------------------------------------------
 """ This example shows spatial interpolation of images. """
@@ -108,7 +106,7 @@ vertices['interpol'] = np.arange(17).reshape(17,1)
 program.bind(vertices)
 indices = np.zeros((17,6),np.uint32).view(gloo.IndexBuffer)
 indices[:] = [0,1,2,1,2,3]
-indices += 4*np.arange(17).reshape(17,1)
+indices += 4*np.arange(17,dtype=np.uint32).reshape(17,1)
 
 lena = data.get("lena.png")
 program['u_data'] = lena
